@@ -1,5 +1,5 @@
 {
-  "name": "candy-drop",
+  "name": "skulla",
   "version": "1.0.0",
   "private": true,
   "dependencies": {
@@ -10,8 +10,8 @@
     "@testing-library/jest-dom": "^5.11.4",
     "@testing-library/react": "^11.1.0",
     "@testing-library/user-event": "^12.1.10",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "react-scripts": "4.0.3",
     "web-vitals": "^1.0.1"
   },
@@ -39,4 +39,5 @@
       "last 1 safari version"
     ]
   }
+  
 }
